+++ android/src/main/java/com/cloudwebrtc/webrtc/MethodCallHandlerImpl.java
@@ -696,6 +696,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
       }
       case "clearAndroidCommunicationDevice": {
         AudioSwitchManager.instance.clearCommunicationDevice();
+        result.success(null);
         break;
       }
       case "setMicrophoneMute":
@@ -769,7 +770,8 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
         break;
       case "stopRecordToFile":
         Integer recorderId = call.argument("recorderId");
-        getUserMediaImpl.stopRecording(recorderId);
+        String albumName = call.argument("albumName");
+        getUserMediaImpl.stopRecording(recorderId, albumName);
         result.success(null);
         break;
       case "captureFrame": {