+++ lib/src/web/rtc_video_renderer_impl.dart
@@ -72,9 +72,6 @@ class RTCVideoRenderer extends ValueNotifier<RTCVideoValue>
   int get videoHeight => value.height.toInt();
 
   @override
-  RTCVideoValue get videoValue => value;
-
-  @override
   int get textureId => _textureId;
 
   @override
@@ -92,12 +89,12 @@ class RTCVideoRenderer extends ValueNotifier<RTCVideoValue>
 
   String get viewType => 'RTCVideoRenderer-$textureId';
 
-  void _updateAllValues() {
-    final element = findHtmlView();
+  void _updateAllValues(web.HTMLVideoElement fallback) {
+    final element = findHtmlView() ?? fallback;
     value = value.copyWith(
       rotation: 0,
-      width: element?.videoWidth.toDouble() ?? 0.0,
-      height: element?.videoHeight.toDouble() ?? 0.0,
+      width: element.videoWidth.toDouble(),
+      height: element.videoHeight.toDouble(),
       renderVideo: renderVideo,
     );
   }
@@ -273,13 +270,13 @@ class RTCVideoRenderer extends ValueNotifier<RTCVideoValue>
 
       _subscriptions.add(
         element.onCanPlay.listen((dynamic _) {
-          _updateAllValues();
+          _updateAllValues(element);
         }),
       );
 
       _subscriptions.add(
         element.onResize.listen((dynamic _) {
-          _updateAllValues();
+          _updateAllValues(element);
           onResize?.call();
         }),
       );