+++ windows/flutter_webrtc_plugin.cc
@@ -4,7 +4,10 @@
 #include "flutter_webrtc.h"
 #include "task_runner_windows.h"
 
+#include <flutter/plugin_registrar_windows.h>
+
 const char* kChannelName = "FlutterWebRTC.Method";
+static flutter_webrtc_plugin::FlutterWebRTC* g_shared_instance = nullptr;
 
 namespace flutter_webrtc_plugin {
 
@@ -21,7 +24,6 @@ class FlutterWebRTCPluginImpl : public FlutterWebRTCPlugin {
     // Uses new instead of make_unique due to private constructor.
     std::unique_ptr<FlutterWebRTCPluginImpl> plugin(
         new FlutterWebRTCPluginImpl(registrar, std::move(channel)));
-
     channel_pointer->SetMethodCallHandler(
         [plugin_pointer = plugin.get()](const auto& call, auto result) {
           plugin_pointer->HandleMethodCall(call, std::move(result));
@@ -47,6 +49,7 @@ class FlutterWebRTCPluginImpl : public FlutterWebRTCPlugin {
         textures_(registrar->texture_registrar()),
         task_runner_(std::make_unique<TaskRunnerWindows>()) {
     webrtc_ = std::make_unique<FlutterWebRTC>(this);
+    g_shared_instance = webrtc_.get();
   }
 
   // Called when a method is called on |channel_|;
@@ -71,7 +74,11 @@ class FlutterWebRTCPluginImpl : public FlutterWebRTCPlugin {
 
 void FlutterWebRTCPluginRegisterWithRegistrar(
     FlutterDesktopPluginRegistrarRef registrar) {
-    static auto* plugin_registrar = new flutter::PluginRegistrar(registrar);
-    flutter_webrtc_plugin::FlutterWebRTCPluginImpl::RegisterWithRegistrar(
-        plugin_registrar);
+  flutter_webrtc_plugin::FlutterWebRTCPluginImpl::RegisterWithRegistrar(
+      flutter::PluginRegistrarManager::GetInstance()
+          ->GetRegistrar<flutter::PluginRegistrarWindows>(registrar));
 }
+
+flutter_webrtc_plugin::FlutterWebRTC* FlutterWebRTCPluginSharedInstance() {
+  return g_shared_instance;
+}